+++ src/components/SelectionGroup/index.js
@@ -1,0 +1,1 @@
+export { default as SelectionGroup } from './SelectionGroup'